Copied over adapter changes from old to new arch
Had to change how the js event was dispatched because new arch threw an error. This is the way to dispatch on new arch - other way doesn't work.
Need to retest this on the old arch
